refactor(mdx-env), move the implementation into MdxEnv class instead of patching it from the main runtime
