+++ scopes/mdx/mdx/mdx.env.ts
@@ -1,7 +1,76 @@
 import { Environment } from '@teambit/envs';
+import { ReactMain } from '@teambit/react';
+import { BabelCompiler } from '@teambit/compilation.babel-compiler';
+import { TypescriptConfigMutator } from '@teambit/typescript.modules.ts-config-mutator';
+import { TsConfigTransformer } from '@teambit/typescript';
+import { babelConfig } from './babel/babel.config';
+import { Logger } from '@teambit/logger';
+import { MultiCompilerMain } from '@teambit/multi-compiler';
+import { CompilerMain } from '@teambit/compiler';
+import { DocsMain } from '@teambit/docs';
+import { MDXAspect } from './mdx.aspect';
+import { MDXCompiler, MDXCompilerOpts } from './mdx.compiler';
 
 export const MdxEnvType = 'mdx';
 
 export class MdxEnv implements Environment {
-  // TODO: add the env here properly.
+  constructor(private react: ReactMain, private logger: Logger,
+    private multiCompiler: MultiCompilerMain, private compiler: CompilerMain, private docs: DocsMain){}
+  getCompiler() {
+    const tsTransformer: TsConfigTransformer = (tsconfig: TypescriptConfigMutator) => {
+      // set the shouldCopyNonSupportedFiles to false since we don't want ts to copy the .mdx file to the dist folder (it will conflict with the .mdx.js file created by the mdx compiler)
+      tsconfig.setCompileJs(false).setCompileJsx(false).setShouldCopyNonSupportedFiles(false);
+      return tsconfig;
+    };
+    const tsCompiler = this.react.env.getCompiler([tsTransformer]);
+
+    const babelCompiler = BabelCompiler.create(
+      {
+        babelTransformOptions: babelConfig,
+        // set the shouldCopyNonSupportedFiles to false since we don't want babel to copy the .mdx file to the dist
+        // folder (it will conflict with the .mdx.js file created by the mdx compiler)
+        shouldCopyNonSupportedFiles: false,
+      },
+      { logger: this.logger }
+    );
+
+
+    return this.multiCompiler.createCompiler(
+      [
+        babelCompiler,
+        this.createMdxCompiler({ ignoredPatterns: this.docs.getPatterns(), babelTransformOptions: babelConfig }),
+        tsCompiler,
+      ],
+      {}
+    );
+  }
+
+  getBuildPipe() {
+    const mdxCompiler = this.getCompiler();
+    return [this.compiler.createTask('MDXCompiler', mdxCompiler), ...this.react.reactEnv.createBuildPipeWithoutCompiler()];
+  }
+
+  async getDependencies() {
+    return {
+      ...this.react.reactEnv.getDependencies(),
+      dependencies: {
+        '@teambit/mdx.ui.mdx-scope-context': '1.0.0',
+        '@mdx-js/react': '1.6.22',
+      },
+    };
+  }
+
+  /**
+   * create an instance of the MDX compiler.
+   */
+  createMdxCompiler(opts: MDXCompilerOpts = {}) {
+    const mdxCompiler = new MDXCompiler(MDXAspect.id, opts);
+    return mdxCompiler;
+  }
+
+  async __getDescriptor() {
+    return {
+      type: 'mdx',
+    };
+  }
 }
